Clean up, moved enum definitions
