+++ gym_pybullet_drones/envs/single_agent_rl/BaseSingleAgentAviary.py
@@ -5,26 +5,9 @@ from gymnasium import spaces
 import pybullet as p
 
 from gym_pybullet_drones.envs.BaseAviary import BaseAviary
-from gym_pybullet_drones.utils.enums import DroneModel, Physics, ImageType
-from gym_pybullet_drones.utils.utils import nnlsRPM
+from gym_pybullet_drones.utils.enums import DroneModel, Physics, ImageType, ActionType, ObservationType
 from gym_pybullet_drones.control.DSLPIDControl import DSLPIDControl
 
-class ActionType(Enum):
-    """Action type enumeration class."""
-    RPM = "rpm"                 # RPMS
-    PID = "pid"                 # PID control
-    VEL = "vel"                 # Velocity input (using PID control)
-    ONE_D_RPM = "one_d_rpm"     # 1D (identical input to all motors) with RPMs
-    ONE_D_PID = "one_d_pid"     # 1D (identical input to all motors) with PID control
-
-################################################################################
-
-class ObservationType(Enum):
-    """Observation type enumeration class."""
-    KIN = "kin"     # Kinematic information (pose, linear and angular velocities)
-    RGB = "rgb"     # RGB camera capture in each drone's POV
-
-################################################################################
 
 class BaseSingleAgentAviary(BaseAviary):
     """Base single drone environment class for reinforcement learning."""
