+++ gym_pybullet_drones/utils/enums.py
@@ -29,3 +29,18 @@ class ImageType(Enum):
     BW = 3      # Black and white
 
 ################################################################################
+
+class ActionType(Enum):
+    """Action type enumeration class."""
+    RPM = "rpm"                 # RPMS
+    PID = "pid"                 # PID control
+    VEL = "vel"                 # Velocity input (using PID control)
+    ONE_D_RPM = "one_d_rpm"     # 1D (identical input to all motors) with RPMs
+    ONE_D_PID = "one_d_pid"     # 1D (identical input to all motors) with PID control
+
+################################################################################
+
+class ObservationType(Enum):
+    """Observation type enumeration class."""
+    KIN = "kin"     # Kinematic information (pose, linear and angular velocities)
+    RGB = "rgb"     # RGB camera capture in each drone's POV
