+++ gym_pybullet_drones/utils/utils.py
@@ -52,82 +52,3 @@ def str2bool(val):
         return False
     else:
         raise argparse.ArgumentTypeError("[ERROR] in str2bool(), a Boolean value is expected")
-
-################################################################################
-
-def nnlsRPM(thrust,
-            x_torque,
-            y_torque,
-            z_torque,
-            counter,
-            max_thrust,
-            max_xy_torque,
-            max_z_torque,
-            a,
-            inv_a,
-            b_coeff,
-            gui=False
-            ):
-    """Non-negative Least Squares (NNLS) RPMs from desired thrust and torques.
-
-    This function uses the NNLS implementation in `scipy.optimize`.
-
-    Parameters
-    ----------
-    thrust : float
-        Desired thrust along the drone's z-axis.
-    x_torque : float
-        Desired drone's x-axis torque.
-    y_torque : float
-        Desired drone's y-axis torque.
-    z_torque : float
-        Desired drone's z-axis torque.
-    counter : int
-        Simulation or control iteration, only used for printouts.
-    max_thrust : float
-        Maximum thrust of the quadcopter.
-    max_xy_torque : float
-        Maximum torque around the x and y axes of the quadcopter.
-    max_z_torque : float
-        Maximum torque around the z axis of the quadcopter.
-    a : ndarray
-        (4, 4)-shaped array of floats containing the motors configuration.
-    inv_a : ndarray
-        (4, 4)-shaped array of floats, inverse of a.
-    b_coeff : ndarray
-        (4,1)-shaped array of floats containing the coefficients to re-scale thrust and torques. 
-    gui : boolean, optional
-        Whether a GUI is active or not, only used for printouts.
-
-    Returns
-    -------
-    ndarray
-        (4,)-shaped array of ints containing the desired RPMs of each propeller.
-
-    """
-    #### Check the feasibility of thrust and torques ###########
-    if gui and thrust < 0 or thrust > max_thrust:
-        print("[WARNING] iter", counter, "in utils.nnlsRPM(), unfeasible thrust {:.2f} outside range [0, {:.2f}]".format(thrust, max_thrust))
-    if gui and np.abs(x_torque) > max_xy_torque:
-        print("[WARNING] iter", counter, "in utils.nnlsRPM(), unfeasible roll torque {:.2f} outside range [{:.2f}, {:.2f}]".format(x_torque, -max_xy_torque, max_xy_torque))
-    if gui and np.abs(y_torque) > max_xy_torque:
-        print("[WARNING] iter", counter, "in utils.nnlsRPM(), unfeasible pitch torque {:.2f} outside range [{:.2f}, {:.2f}]".format(y_torque, -max_xy_torque, max_xy_torque))
-    if gui and np.abs(z_torque) > max_z_torque:
-        print("[WARNING] iter", counter, "in utils.nnlsRPM(), unfeasible yaw torque {:.2f} outside range [{:.2f}, {:.2f}]".format(z_torque, -max_z_torque, max_z_torque))
-    B = np.multiply(np.array([thrust, x_torque, y_torque, z_torque]), b_coeff)
-    sq_rpm = np.dot(inv_a, B)
-    #### NNLS if any of the desired ang vel is negative ########
-    if np.min(sq_rpm) < 0:
-        sol, res = nnls(a,
-                        B,
-                        maxiter=3*a.shape[1]
-                        )
-        if gui:
-            print("[WARNING] iter", counter, "in utils.nnlsRPM(), unfeasible squared rotor speeds, using NNLS")
-            print("Negative sq. rotor speeds:\t [{:.2f}, {:.2f}, {:.2f}, {:.2f}]".format(sq_rpm[0], sq_rpm[1], sq_rpm[2], sq_rpm[3]),
-                   "\t\tNormalized: [{:.2f}, {:.2f}, {:.2f}, {:.2f}]".format(sq_rpm[0]/np.linalg.norm(sq_rpm), sq_rpm[1]/np.linalg.norm(sq_rpm), sq_rpm[2]/np.linalg.norm(sq_rpm), sq_rpm[3]/np.linalg.norm(sq_rpm)))
-            print("NNLS:\t\t\t\t [{:.2f}, {:.2f}, {:.2f}, {:.2f}]".format(sol[0], sol[1], sol[2], sol[3]),
-                  "\t\t\tNormalized: [{:.2f}, {:.2f}, {:.2f}, {:.2f}]".format(sol[0]/np.linalg.norm(sol), sol[1]/np.linalg.norm(sol), sol[2]/np.linalg.norm(sol), sol[3]/np.linalg.norm(sol)),
-                  "\t\tResidual: {:.2f}".format(res))
-        sq_rpm = sol
-    return np.sqrt(sq_rpm)
